pyproject.toml
[tool.poetry]
name = "elftool"
version = "0.1.0"
description = ""
authors = ["Sergey Silaev <hello@8am.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyelftools = "^0.29"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.282"
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
